+++ src/Consumer/Reader/MultiQueueReader.php
@@ -2,6 +2,7 @@
 
 namespace Bdf\Queue\Consumer\Reader;
 
+use Bdf\Queue\Connection\ConnectionDriverInterface;
 use Bdf\Queue\Connection\QueueDriverInterface;
 use Bdf\Queue\Message\EnvelopeInterface;
 
@@ -60,5 +61,13 @@ final class MultiQueueReader implements QueueReaderInterface
     public function stop(): void
     {
         $this->driver->connection()->close();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function connection(): ConnectionDriverInterface
+    {
+        return $this->driver->connection();
     }
 }