+++ src/Consumer/Reader/QueueReader.php
@@ -2,6 +2,7 @@
 
 namespace Bdf\Queue\Consumer\Reader;
 
+use Bdf\Queue\Connection\ConnectionDriverInterface;
 use Bdf\Queue\Connection\QueueDriverInterface;
 use Bdf\Queue\Message\EnvelopeInterface;
 
@@ -47,5 +48,13 @@ final class QueueReader implements QueueReaderInterface
     public function stop(): void
     {
         $this->driver->connection()->close();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function connection(): ConnectionDriverInterface
+    {
+        return $this->driver->connection();
     }
 }