+++ tests/Consumer/Receiver/Builder/ReceiverBuilderTest.php
@@ -491,8 +491,8 @@ class MyReceiver implements ReceiverInterface
     public function start(ConsumerInterface $consumer): void {}
     public function receive($message, ConsumerInterface $consumer): void {}
     public function receiveTimeout(ConsumerInterface $consumer): void {}
-    public function receiveStop(): void {}
-    public function terminate(): void {}
+    public function receiveStop(ConsumerInterface $consumer): void {}
+    public function terminate(ConsumerInterface $consumer): void {}
 }
 
 class MyHandler